refactor: move local storage related func to 'services/'
