--- project/src/utils/token.ts
@@ -1,16 +1,0 @@
-export type Token = string;
-
-const AUTH_TOKEN_KEY_NAME = 'htmlacademy-794021-six-sities-simple';
-
-export const getToken = (): Token => {
-  const token = localStorage.getItem(AUTH_TOKEN_KEY_NAME);
-  return token ?? '';
-};
-
-export const saveToken = (token: Token): void => {
-  localStorage.setItem(AUTH_TOKEN_KEY_NAME, token);
-};
-
-export const dropToken = (): void => {
-  localStorage.removeItem(AUTH_TOKEN_KEY_NAME);
-};
